# frontend/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: ILMA API
  version: 1.0.0
  description: |
    API for ILMA Education Platform (Offline-First).
    Targets CM2 students in Benin preparing for CEP exam.
    Three roles: Student, Parent, Admin.
servers:
  - url: https://api.ilma.app/v1
    description: Production
  - url: https://staging-api.ilma.app/v1
    description: Staging
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum: [VALIDATION_ERROR, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, SERVER_ERROR, NETWORK_ERROR, TIMEOUT, UNKNOWN]
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    User:
      type: object
      required: [id, name, email, role]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        email: { type: string, format: email }
        phone: { type: string }
        role: { type: string, enum: [STUDENT, PARENT, ADMIN, GUEST] }
        avatarUrl: { type: string, format: uri }
        level: { type: integer }
        xp: { type: integer }
        xpToNextLevel: { type: integer }
        streak: { type: integer }
        subscriptionTier: { type: string, enum: [FREE, PREMIUM] }
    AuthResponse:
      type: object
      required: [accessToken, user]
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        expiresIn: { type: integer }
        user: { $ref: '#/components/schemas/User' }
...
    Notification:
      type: object
      properties:
        id: { type: string }
        type: { type: string, enum: [STREAK_DANGER, BADGE_EARNED, WEEKLY_REPORT, GOAL_REACHED, INACTIVITY, SYSTEM] }
        title: { type: string }
        message: { type: string }
        read: { type: boolean }
        createdAt: { type: string, format: date-time }
...
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Login with email/phone + password or OTP
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                identifier: { type: string, description: "Email or phone number" }
                password: { type: string }
                otp: { type: string }
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, role]
              properties:
                name: { type: string }
                email: { type: string }
                phone: { type: string }
                password: { type: string }
                role: { type: string, enum: [STUDENT, PARENT] }
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '400':
          description: Validation error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /auth/otp/send:
    post:
      tags: [Auth]
      summary: Send OTP code
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier]
              properties:
                identifier: { type: string }
                channel: { type: string, enum: [SMS, EMAIL] }
      responses:
        '200': { description: OTP sent }
  /auth/otp/verify:
    post:
      tags: [Auth]
      summary: Verify OTP code
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier, code]
              properties:
                identifier: { type: string }
                code: { type: string }
      responses:
        '200':
          description: OTP verified
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
...
  /notifications:
    get:
      tags: [Notifications]
      summary: List notifications
      parameters:
        - name: unreadOnly
          in: query
          schema: { type: boolean }
      responses:
        '200':
          description: Notifications list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Notification' }
  /notifications/{id}/read:
    put:
      tags: [Notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204': { description: Marked as read }

  # --- SUBSCRIPTION ---
...
  /packs/{packId}/download:
    get:
      tags: [Content Packs]
      summary: Download a content pack
      parameters:
        - name: packId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Pack data (binary)
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary

  # --- ADMIN ---
...